我又又不会做以前做过的题了

+++ Algorithm/CF/Educational Codeforces Round 31/D redo.cpp
@@ -32,9 +32,9 @@ int main()
     LL penal = 0;
     while(pq.size()>=3)
     {
-        int a = pq.top();pq.pop();
-        int b = pq.top();pq.pop();
-        int c = pq.top();pq.pop();
+        LL a = pq.top();pq.pop();
+        LL b = pq.top();pq.pop();
+        LL c = pq.top();pq.pop();
         pq.push(a+b+c);
         penal +=a+b+c;
     }
